Do the sorting in the model
