+++ src/main/java/com/github/alexfalappa/nbspringboot/navigator/ElementScanningTaskFactory.java
@@ -17,9 +17,6 @@ package com.github.alexfalappa.nbspringboot.navigator;
 
 import com.sun.source.tree.CompilationUnitTree;
 import com.sun.source.util.TreePath;
-import java.util.Collections;
-import java.util.Comparator;
-import java.util.List;
 import org.netbeans.api.java.source.CancellableTask;
 import org.netbeans.api.java.source.CompilationInfo;
 import org.netbeans.api.java.source.JavaSource;
@@ -69,25 +66,7 @@ public class ElementScanningTaskFactory extends LookupBasedJavaSourceTaskFactory
             final TreePath rootPath = new TreePath(compilationUnitTree);
 
             this.mappedElementExtractor = new MappedElementExtractor(p.getFileObject(), compilationUnitTree, p.getTrees(), rootPath);
-            final List<MappedElement> mappedElements = compilationUnitTree.accept(this.mappedElementExtractor, null);
-            Collections.sort(mappedElements, new Comparator<MappedElement>() {
-                @Override
-                public int compare(MappedElement o1, MappedElement o2) {
-                    int rv = o1.getResourceUrl().compareTo(o2.getResourceUrl());
-                    if (rv == 0) {
-                        if (o1.getRequestMethod() == null) {
-                            rv = -1;
-                        } else if (o2.getRequestMethod() == null) {
-                            rv = 1;
-                        } else {
-                            rv = o1.getRequestMethod().compareTo(o2.getRequestMethod());
-                        }
-
-                    }
-                    return rv;
-                }
-            });
-            this.targetModel.refresh(mappedElements);
+            this.targetModel.refresh(compilationUnitTree.accept(this.mappedElementExtractor, null));
         }
     }
